Added script to fix d.ts files.
