+++ .github/workflows/release-test.yml
@@ -42,6 +42,8 @@ jobs:
         run: npx nx run-many --target=build --all --parallel=2 --exclude testfront,storybook,fulfillment,storefront
       - name: Fix js extension
         run: node tools/post-build/add-js-extension.js
+      - name: Fix d.ts files
+        run: node tools/post-build/fix-dts-paths.js
       - name: Install verdaccio
         run: npm i -g verdaccio
       - name: Run verdaccio first time to generate config
